Add generics tests
