+++ goose.go
@@ -2616,16 +2616,11 @@ func (ctx *Ctx) funcDecl(d *ast.FuncDecl) []glang.Decl {
 	defer ctx.dep.UnsetCurrentName()
 
 	if d.Type.TypeParams != nil {
-		// TODO(generics): a plan for supporting generics:
-		// - embed `go_type` into GooseLang rather than just Gallina; this
-		//   includes instantiated generic types which includes the type
-		//   arguments.
-		// - type parameters would be GooseLang arguments, and method lookups
-		//   would look up based on the type name and then specialize the method.
-		// - interface values need to remember if the type is generic, and the
-		//   type arguments if yes, so the type arguments can be passed into the
-		//   method from `interface.get`
-		ctx.unsupported(d, "generic functions not supported")
+		for _, p := range d.Type.TypeParams.List {
+			for _, name := range p.Names {
+				fd.TypeArgs = append(fd.TypeArgs, glang.Binder{Name: name.Name})
+			}
+		}
 	}
 
 	if d.Body == nil {
